Extracted JacobiKey from PolynomialsUtils.
Fixes #322
